Added sprite renderer; Added sprite shader; Added dev window; Added wireframe; Fixed orthographic camera implementation.

+++ StudentEngine/src/graphics/buffers/vertexArray.h
@@ -14,12 +14,12 @@ public:
 		uint32 attributeIndex = 0;
 		for (int i = 0; i < m_buffers.size(); i++) {
 			m_buffers[i]->ApplyLayout(attributeIndex);
-			attributeIndex += m_buffers[i]->GetLayout().GetElements().size();
+			attributeIndex += (uint32)m_buffers[i]->GetLayout().GetElements().size();
 		}
 		Unbind();
 	}
 
-	int GetBufferCount() { return m_buffers.size(); }
+	int GetBufferCount() { return (int)m_buffers.size(); }
 	vector<ManagedRef<VertexBuffer>> GetBuffers() { return m_buffers; }
 	ManagedRef<VertexBuffer> GetBuffer(int index) { return m_buffers.at(index); }
 	void AddBuffer(ManagedRef<VertexBuffer> buffer);
